Equals/HashCode refactor
As pretty much all the classes were relying on mutable data for
equals/hashcode which is not ideal. Though hashcode was being fixed it
could still lead to equals returning false when it should be true.

As such Equals and HashCode returned to defaults in the main, with any
equality checks involving the mutable data moved to another method
(isEqualTo), default equals simply checks they are the same object and
HashSets etc. should work normally